+++ core/models/__init__.py
@@ -44,6 +44,9 @@ from core.models.capability_readiness import (
     CapabilityReadinessRecord as CapabilityReadinessRecord,
 )
 from core.models.case_pseudonym_map import CasePseudonymMap as CasePseudonymMap
+from core.models.case_push import CasePushEndpoint as CasePushEndpoint
+from core.models.case_push import CasePushOutbox as CasePushOutbox
+from core.models.case_push import ProviderWebhookReceipt as ProviderWebhookReceipt
 from core.models.cdc import CDCEvent as CDCEvent
 from core.models.cdc import CDCEventDeadLetter as CDCEventDeadLetter
 from core.models.client_portal import ClientPortalDocument as ClientPortalDocument